+++ pgf_kernel_experiments/experiments/earth_topography/tune_multiple_gps.py
@@ -151,11 +151,11 @@ ax[3].set_yticks(ticks=ytick_ticks, labels=ytick_labels, fontsize=fontsize)
 
 # %% Convert training and test data to PyTorch format
 
-train_x = torch.as_tensor(train_pos, dtype=torch.float32)
-train_y = torch.as_tensor(train_output, dtype=torch.float32)
+train_x = torch.as_tensor(train_pos, dtype=torch.float64)
+train_y = torch.as_tensor(train_output, dtype=torch.float64)
 
-test_x = torch.as_tensor(test_pos, dtype=torch.float32)
-test_y = torch.as_tensor(test_output, dtype=torch.float32)
+test_x = torch.as_tensor(test_pos, dtype=torch.float64)
+test_y = torch.as_tensor(test_output, dtype=torch.float64)
 
 # %% Set up ExactMultiGPRunner
