+++ pgf_kernel_experiments/experiments/gaussian_srf/tune_multiple_gps.py
@@ -86,11 +86,11 @@ ax[3].set_yticks(np.linspace(-1, 1, num=5), fontsize=fontsize)
 
 # %% Convert training and test data to PyTorch format
 
-train_x = torch.as_tensor(train_pos.T, dtype=torch.float32)
-train_y = torch.as_tensor(train_output.T, dtype=torch.float32)
+train_x = torch.as_tensor(train_pos.T, dtype=torch.float64)
+train_y = torch.as_tensor(train_output.T, dtype=torch.float64)
 
-test_x = torch.as_tensor(test_pos.T, dtype=torch.float32)
-test_y = torch.as_tensor(test_output.T, dtype=torch.float32)
+test_x = torch.as_tensor(test_pos.T, dtype=torch.float64)
+test_y = torch.as_tensor(test_output.T, dtype=torch.float64)
 
 # %% Set up ExactMultiGPRunner
